Attempt to drain controller output on POLLHUP
Testing on Linux 3.13 indicates that, if a process connected to a pty
slave writes to the terminal and terminates, Python epoll sometimes
returns select.POLLHUP prior to select.POLLIN on the pty master.  This
out-of-order delivery may cause Ashier to drop controller keystrokes.

This commit works around the problem by copying data from the controller
to the child pty when Ashier receives select.POLLHUP on the controller
pty.  If epoll happens to deliver updates in order, this final copy
attempt will fail, so this commit also makes CopyData swallow all
OSError exceptions.